requirements.txt
# Python 3.12 Compatibility Dependencies for f110_gym
# Use pip install -r requirements.txt to install these packages

# Core Libraries
numpy>=1.23.0
scipy
pyyaml

# Reinforcement Learning Environment API
gymnasium>=1.2.0

# Performance (JIT Compilation)
numba>=0.61.0

# Visualization and Rendering
pyglet>=2.1.0
Pillow

# Optional/Utility (Include if your code uses them)
shapely
matplotlib
